02_NLP/app/PictureBooks/requirements.txt
reportlab
openai
Pillow